Fix templates for VS for Mac 2022 (#7764)
* Fix templates for VS for Mac 2022

* Restore groupIdentitiy to templates

* Hook up VSMacExtension to the build system

* Fix CAKE script

* We can't build VS for Mac 2022 addin on github actions until it goes stable
