+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.Application.UWP.XAML.CSharp/.template.config/template.json
@@ -4,6 +4,7 @@
         "MonoGame", "Games", "Desktop", "Windows", "Xbox", "UWP", "XAML"
     ],
     "name": "MonoGame Windows Universal XAML Application",
+    "groupIdentity": "MonoGame.Application.UWP.XAML",
     "identity": "MonoGame.Application.UWP.XAML.CSharp",
     "shortName": "mguwpxaml",
     "tags": {
