+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.Application.WindowsDX.CSharp/.template.config/template.json
@@ -4,6 +4,7 @@
         "MonoGame", "Games", "Desktop", "Windows", "Linux", "macOS"
     ],
     "name": "MonoGame Windows Desktop Application",
+    "groupIdentity": "MonoGame.Application.WindowsDX",
     "identity": "MonoGame.Application.WindowsDX.CSharp",
     "shortName": "mgwindowsdx",
     "tags": {
