+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.Application.iOS.CSharp/.template.config/template.json
@@ -4,6 +4,7 @@
         "MonoGame", "Games", "Mobile", "iOS"
     ],
     "name": "MonoGame iOS Application",
+    "groupIdentity": "MonoGame.Application.iOS",
     "identity": "MonoGame.Application.iOS.CSharp",
     "shortName": "mgios",
     "tags": {
