+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.Library.CSharp/.template.config/template.json
@@ -4,6 +4,7 @@
         "MonoGame", "Games", "Library"
     ],
     "name": "MonoGame Game Library",
+    "groupIdentity": "MonoGame.Library",
     "identity": "MonoGame.Library.CSharp",
     "shortName": "mglib",
     "tags": {
