+++ Templates/MonoGame.Templates.CSharp/content/MonoGame.Library.Pipeline.Extension.CSharp/.template.config/template.json
@@ -4,6 +4,7 @@
         "MonoGame", "Games", "Extensions"
     ],
     "name": "MonoGame Content Pipeline Extension",
+    "groupIdentity": "MonoGame.Library.Pipeline",
     "identity": "MonoGame.Library.Pipeline.Extension.CSharp",
     "shortName": "mgpipeline",
     "tags": {
