+++ Templates/MonoGame.Templates.VSMacExtension/Icons/monogame-16@2x.png
