+++ Templates/MonoGame.Templates.VSMacExtension/Icons/monogame-16~dark@2x.png
