+++ Templates/MonoGame.Templates.VSMacExtension/Icons/monogame-16~dark~sel.png
