+++ Templates/MonoGame.Templates.VSMacExtension/Icons/monogame-16~sel.png
