+++ Templates/MonoGame.Templates.VSMacExtension/Icons/monogame-project-32.png
