+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-32~dark@2x.png
