+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-32~dark~sel.png
