+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-32~sel@2x.png
