+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-template@2x.png
