+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-template~dark.png
