+++ Templates/MonoGame.Templates.VSMacExtension/Icons/project-monogame-template~dark@2x.png
