+++ Templates/MonoGame.Templates.VSMacExtension/MonoGame.Templates.VSMacExtension.csproj
@@ -1,0 +1,44 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <BaseOutputPath>..\..\Artifacts\VisualStudioForMac</BaseOutputPath>
+    <CreatePackage>true</CreatePackage>
+    <ImplicitUsings>true</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.VisualStudioMac.Sdk" Version="17.0.0-preview.7955" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Remove="Icons/*.png" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Include="Icons/*.png">
+      <LogicalName>%(Filename)%(Extension)</LogicalName>
+    </EmbeddedResource>
+  </ItemGroup>
+
+  <ItemGroup>
+    <AddinFile Include="..\..\Artifacts\MonoGame.Templates.CSharp\Release\*.nupkg">
+      <Link>Templates\MonoGame.Templates.CSharp.nupkg</Link>
+    </AddinFile>
+  </ItemGroup>
+
+  <Target Name="_GenerateAddinInfo" BeforeTargets="CoreCompile">
+     <ItemGroup>
+       <AddinInfo Include="using System%3B" />
+       <AddinInfo Include="using Mono.Addins%3B" />
+       <AddinInfo Include="using Mono.Addins.Description%3B" />
+       <AddinInfo Include="[assembly: Addin(&quot;MonoGame.Templates.VSMacExtension&quot;, Version = &quot;$(Version)&quot;)]" />
+       <AddinInfo Include="[assembly: AddinName(&quot;MonoGame Extension&quot;)]" />
+       <AddinInfo Include="[assembly: AddinCategory(&quot;Game Development&quot;)]" />
+       <AddinInfo Include="[assembly: AddinDescription(&quot;MonoGame Extension ofr VisualStudio for Mac&quot;)]" />
+       <AddinInfo Include="[assembly: AddinAuthor(&quot;MonoGame Team&quot;)]" />
+     </ItemGroup>
+     <WriteLinesToFile File="Properties\AddinInfo.cs" Lines="@(AddinInfo)" Overwrite="true" />
+  </Target>
+
+</Project>
