+++ Templates/MonoGame.Templates.VSMacExtension/Properties/AddinInfo.cs
@@ -1,0 +1,8 @@
+using System;
+using Mono.Addins;
+using Mono.Addins.Description;
+[assembly: Addin("MonoGame.Templates.VSMacExtension", Version = "3.8.1.137-develop")]
+[assembly: AddinName("MonoGame Extension")]
+[assembly: AddinCategory("Game Development")]
+[assembly: AddinDescription("MonoGame Extension ofr VisualStudio for Mac")]
+[assembly: AddinAuthor("MonoGame Team")]
