+++ Templates/MonoGame.Templates.VSMacExtension/Properties/Manifest.addin.xml
@@ -1,7 +1,9 @@
 ﻿<?xml version="1.0" encoding="UTF-8"?>
 <ExtensionModel>
+
 	<Runtime>
 	</Runtime>
+
 	<Extension path="/MonoDevelop/Ide/ProjectTemplateCategories">
 		<Category id="monogame" _name="MonoGame" icon="monogame-project-icon-16" insertbefore="other">
 			<Category id="app" _name="App">
@@ -12,6 +14,7 @@
 			</Category>
 		</Category>
 	</Extension>
+
 	<Extension path="/MonoDevelop/Ide/Templates">
 		<Template
 			id="MonoGame.Application.DesktopGL.CSharp"
@@ -63,30 +66,22 @@
 			_overrideDescription="A MonoGame Content Pipeline Extensions project. Use this to extend the MonoGame Content Pipleine with your own Importers or Processors."
 			category="monogame/library/general"/>
 	</Extension>
+
 	<Extension path="/MonoDevelop/Ide/TemplateImages">
 		<Image id="md-monogame-project" resource="monogame-project-32.png" />
 		<Image id="md-monogame-icon-project" resource="monogame-project-32.png" />
 		<Image id="md-monogame-template" resource="project-monogame-template.png" />
-		<Image id="md-monogame-android-project" resource="monogame-android.png" />
-		<Image id="md-monogame-ios-project" resource="monogame-ios.png" />
-		<Image id="md-monogame-desktopgl-project" resource="monogame-desktopgl.png" />
-		<Image id="md-monogame-windows-project" resource="monogame-windows.png" />
-		<Image id="md-monogame-shared-project" resource="monogame-shared.png" />
 		<Image id="md-monogame-preview-project" resource="monogame-preview.png" />
 	</Extension>
+
 	<Extension path="/MonoDevelop/Core/StockIcons">
 		<StockIcon stockid="monogame-project-icon-16" resource="monogame-16.png" size="Dnd" />
 		<StockIcon stockid="monogame-project-icon" resource="monogame-project-32.png" size="Dnd" />
 		<StockIcon stockid="md-monogame-project" resource="project-monogame-32.png" size="Dnd" />
 		<StockIcon stockid="md-monogame-template" resource="project-monogame-template.png" size="Dnd" />
 		<StockIcon stockid="md-monogame-icon-project" resource="project-monogame-32.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-android-project" resource="monogame-android.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-ios-project" resource="monogame-ios.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-desktopgl-project" resource="monogame-desktopgl.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-windows-project" resource="monogame-windows.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-shared-project" resource="monogame-shared.png" size="Dnd" />
-		<StockIcon stockid="md-monogame-preview-project" resource="monogame-preview.png" size="Dnd" />
 	</Extension>
+
 	<Extension path="/MonoDevelop/Core/MimeTypes">
 		<MimeType id="x-mgcb" _description="MonoGame Content Build File" icon="monogame-project-icon" isText="false">
 			<File pattern="*.mgcb" />
@@ -95,11 +90,9 @@
             <File pattern="*.spritefont" />
         </MimeType>
 	</Extension>
+
 	<Extension path="/MonoDevelop/Ide/DisplayBindings">
-        <DisplayBinding id="MGCBPadBinding" insertbefore="DefaultDisplayBinding" class="MonoGame.IDE.VisualStudioForMac.MGCBPadBinding" />
-		<DisplayBinding id="MGCBEditorBinding" insertbefore="DefaultDisplayBinding" class="MonoGame.IDE.VisualStudioForMac.MGCBEditorBinding" />
+		<DisplayBinding id="MGCBEditorDisplayBinding" insertbefore="DefaultDisplayBinding" class="MonoGame.Templates.VSMacExtension.MGCBEditorDisplayBinding" />
 	</Extension>
-    <Extension path = "/MonoDevelop/Ide/Pads">
-       <Pad id = "MGPad" _label = "MonoGame Content" icon="monogame-project-icon-16" group="MonoGame" class="MonoGame.IDE.VisualStudioForMac.MGPad" defaultPlacement="Left" defaultLayout="*"/>
-    </Extension>
+
 </ExtensionModel>
