--- Templates/VisualStudioForMac/Properties/AddinInfo.cs
@@ -1,8 +1,0 @@
-using System;
-using Mono.Addins;
-using Mono.Addins.Description;
-[assembly: Addin("MonoGame_IDE_VisualStudioForMac",Namespace = "MonoDevelop",Version = "3.8.0.1")]
-[assembly: AddinName("MonoGame Extension")]
-[assembly: AddinCategory("Game Development")]
-[assembly: AddinDescription("VisualStudio for Mac extension for MonoGame")]
-[assembly: AddinAuthor("@MonoGameTeam")]
